% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBreakpoints}
\alias{plotBreakpoints}
\title{Plotting function for BreakPointR}
\usage{
plotBreakpoints(data, file = NULL)
}
\arguments{
\item{output}{genomic ranges from the function runBreakpointr or runBreakpointrALL}
}
\description{
Plotting function for BreakPointR
}
\author{
David Porubsky
}

