% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeBedFile.R
\name{writeBedFile}
\alias{writeBedFile}
\title{Generates a bedfile from inFile}
\usage{
writeBedFile(fileName, dataDirectory, inFile, deltaGraph = F,
  breakTrack = NULL, bin = "NA", summary = F)
}
\arguments{
\item{fileName}{is used to index the bedfile(s)}

\item{inFile}{A \code{\link{GRanges}} object with strand and mapq metadata, such as that generated by \code{\link{bam2GRanges}}}

\item{deltaGraph}{If \code{TRUE}, will generate a bedgraph from a \code{\link{GRanges}} object with metadata column "deltaW" generated by \code{\link{deltaWCalculator}}}

\item{breakTrack}{If given a \code{\link{GRanges}} object with metadata "genoT" (e.g. newBreaks) will write a bedtrack with refined breakpoints}

\item{summary}{If \code{TRUE} appends these data to a file for each index (with all chromosomes included)}

\item{fileDestination}{is location to write bedfile(s)}
}
\description{
Write a bedfile from Breakpoint.R files for upload on to UCSC Genome browser
}
\author{
Ashley Sanders, David Porubsky, Aaron Taudt
}

